updated ssz sizes to use long

+++ infrastructure/ssz/src/main/java/tech/pegasys/teku/infrastructure/ssz/schema/SszUnionSchema.java
@@ -54,7 +54,7 @@ public interface SszUnionSchema<SszUnionT extends SszUnion> extends SszSchema<Ss
   }
 
   @Override
-  default int getSszFixedPartSize() {
+  default long getSszFixedPartSize() {
     return SELECTOR_SIZE_BYTES;
   }
 }
